requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.3
altair==4.2.2
amqp==5.1.1
aniso8601==9.0.1
annotated-types==0.6.0
ansicon==1.89.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.6.0
astor==0.8.1
asttokens==2.2.1
async-lru==2.0.4
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
billiard==4.2.0
bleach==6.1.0
blessed==1.20.0
blinker==1.6.3
cachelib==0.9.0
cachetools==5.3.0
celery==5.3.5
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.0.1
checksum==1.0.4
checksumdir==1.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.1.3
comtypes==1.2.0
contourpy==1.0.7
cors==1.0.1
cycler==0.11.0
Cython==0.29.34
dacite==1.8.1
dataclasses-json==0.5.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==4.2.1
dnspython==2.3.0
docopt==0.6.2
dominate==2.9.1
editor==1.6.6
email-validator==2.0.0.post2
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.93.0
fastjsonschema==2.19.1
ffmpeg-python==0.2.0
ffmpy==0.3.0
filelock==3.9.0
flasgger==0.9.5
Flask==2.3.2
Flask-Alembic==2.0.1
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.1.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-OAuthlib==0.9.6
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Security==3.0.0
Flask-Security-Too==5.0.2
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
Flask-Uploads==0.2.1
Flask-WTF==1.1.1
fonttools==4.39.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
gevent==22.10.2
git-python==1.0.3
gitdb==4.0.10
GitPython==3.1.31
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.135.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.1
googleapis-common-protos==1.63.2
googletrans==3.0.0
gradio==3.20.1
graphviz==0.20.3
greenlet==2.0.2
grpcio==1.64.1
grpcio-status==1.62.2
gTTS==2.3.1
gunicorn==21.2.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
html2image==2.0.4.3
htmlmin==0.1.12
httpcore==1.0.2
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.16.4
hyperframe==5.2.0
idna==2.10
ImageHash==4.3.1
importlib-metadata==7.0.1
inquirer==3.2.3
ipykernel==6.29.0
ipython==8.12.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jinxed==1.2.1
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.0
jupyter_core==5.3.0
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
kiwisolver==1.4.4
kombu==5.3.4
langchain==0.0.155
linkify-it-py==2.0.0
litellm==1.22.3
llvmlite==0.40.0
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
missingno==0.5.2
mistune==2.0.5
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
multimethod==1.10
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.5.6
networkx==3.1
networkz==1.0.6
notebook==7.2.0
notebook_shim==0.2.4
numba==0.57.0
numexpr==2.8.4
numpy==1.24.2
oauthlib==2.1.0
ollama==0.2.0
open-interpreter==0.2.0
openai==0.27.1
openai-whisper==20230314
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
orjson==3.8.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.0
pandas==1.5.3
pandas-profiling==3.2.0
pandocfilters==1.5.1
parso==0.8.3
passlib==1.7.4
patsy==0.5.4
phik==0.12.3
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.4.13
platformdirs==3.2.0
pocketsphinx==5.0.3
posthog==3.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.38
proto-plus==1.24.0
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.13
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.15
pydantic_core==2.14.5
pydeck==0.8.1b0
pydot==2.0.0
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.8.0
Pympler==1.0.1
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
pyparsing==3.0.9
PyPDF2==3.0.1
pypiwin32==223
pyreadline3==3.4.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.6
pyttsx3==2.90
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.5.0
pywin32==305
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==25.0.2
qtconsole==5.5.2
QtPy==2.4.1
readchar==4.0.5
redis==4.6.0
referencing==0.35.1
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.1.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.18.1
rsa==4.9
runs==1.2.2
safetensors==0.3.1
scikit-learn==1.5.1
scipy==1.11.4
seaborn==0.12.2
selenium==4.17.2
semantic-version==2.10.0
Send2Trash==1.8.3
setuptools-rust==1.6.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.5
speaklater==1.3
SpeechRecognition==3.10.0
SQLAlchemy==2.0.21
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.25.0
statsmodels==0.14.0
streamlit==1.22.0
sympy==1.12
tangled-up-in-unicode==0.2.0
tenacity==8.2.2
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.4.0
tinycss2==1.3.0
tldextract==3.4.0
tokenizers==0.13.3
tokentrim==0.1.13
toml==0.10.2
toolz==0.12.0
torch==2.0.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.30.2
translator==0.0.9
trio==0.24.0
trio-websocket==0.11.1
typeguard==4.1.5
types-python-dateutil==2.9.0.20240316
typing-inspect==0.8.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==4.3
uc-micro-py==1.0.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
validators==0.20.0
vine==5.1.0
visions==0.7.5
visitor==0.1.3
watchdog==3.0.0
Wave==0.0.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==10.4
Werkzeug==2.3.6
wget==3.2
widgetsnbextension==4.0.9
wincertstore==0.2
wordcloud==1.9.3
wsproto==1.2.0
WTForms==3.0.1
xmod==1.8.1
yarg==0.1.9
yarl==1.8.2
ydata-profiling==4.6.3
zipp==3.15.0
zope.event==4.6
zope.interface==5.5.2
